refactor airflow code
